backend/foodgram/api/permissions.py
from rest_framework import permissions


class AuthenticatedOrAuthorOrReadOnly(permissions.IsAuthenticatedOrReadOnly):
    def has_object_permission(self, request, view, obj):
        return (request.method in permissions.SAFE_METHODS
                or request.user == obj.author)
