backend/foodgram/requirements.txt
Django==3.2
djangorestframework==3.12.4
psycopg2-binary==2.9.6
djoser==2.1.0
Pillow==9.5.0
drf-extra-fields==3.5.0
gunicorn==20.0.4
